Chapter Module done

+++ app/Services/CategoryService.php
@@ -79,7 +79,10 @@ class CategoryService
 
         if($request->thumbnail){
             $thumbnail = $this->fileUpload($request, 'thumbnail', 'class_thumbnail');
-            $this->deleteFile($category->thumbnail);
+            if($category->thumbnail){
+                $this->deleteFile($category->thumbnail);
+            }
+            
             $category_data = [
                 'name_bn' => $request->name_bn,
                 'name_en' => $request->name_en,
@@ -97,7 +100,9 @@ class CategoryService
     {
         try {
             $category = Category::findOrFail($id);
-            $this->deleteFile($category->thumbnail);
+            if($category->thumbnail){
+                $this->deleteFile($category->thumbnail);
+            }
 
             return Category::findOrFail($id)->delete();
         } catch (\Exception $e) {
